Added metadata logging and load balancing between chunks. Currently the metadata is backed up to the database in a clunky way, see comments. The server can now recover from failures (except if it fails at the extremely incovenient moment of backing up the metadata). Chunk ports curently need to be passed in to the server via command line arguments in the following format (port1, port2, port3 ...)

+++ src/main/java/server/FileSystemImpl.java
@@ -1,42 +1,25 @@
 package server;
 
-import static java.util.Arrays.asList;
-
 import chunk.Chunk;
-import com.mongodb.client.MongoClient;
-import com.mongodb.client.MongoClients;
-import com.mongodb.client.MongoCollection;
-import com.mongodb.client.MongoDatabase;
-import com.mongodb.client.model.Filters;
-import com.mongodb.client.model.Updates;
-import com.mongodb.client.result.DeleteResult;
+import com.mongodb.client.*;
 import org.apache.thrift.TApplicationException;
 import org.apache.thrift.TException;
 
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
-import java.util.logging.Filter;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 import Helper.Logger;
 import org.apache.thrift.protocol.TBinaryProtocol;
 import org.apache.thrift.protocol.TProtocol;
 import org.apache.thrift.transport.TSocket;
 import org.apache.thrift.transport.TTransport;
-import org.bson.BsonBinary;
 import org.bson.Document;
-import org.bson.conversions.Bson;
 import org.bson.types.Binary;
 import org.bson.types.ObjectId;
-
-import javax.print.Doc;
 
 /**
  *
@@ -53,27 +36,80 @@ public class FileSystemImpl implements FileSystem.Iface {
      */
     private Chunk.Client chunkClient;
 
-
     private String hostName;
+
+    /**
+     *  fileLocator keeps track of file names and the port number of which chunk they reside in
+     *  loadTracker keeps track of chunk port numbers and the amount of files the chunks contain.
+     */
+    private Map<String, Integer> fileLocator;
+    private Map<Integer, Integer> loadTracker;
 
     /**
      * Creates a new instance of the FileSystemImpl
      *
      * @param logger The logger for logging.
      */
-    public FileSystemImpl(Logger logger) {
+    public FileSystemImpl(Logger logger, ArrayList<Integer> chunkPorts) {
         this.logger = logger;
+        this.fileLocator = new ConcurrentHashMap<>();
+        this.loadTracker = new ConcurrentHashMap<>();
+        for (int port : chunkPorts){
+            this.loadTracker.put(port, 0);
+        }
+        retrieveMetadata();
+    }
+
+    // TODO IMPORTANT: improve these two methods to not delete and recreate metadata but rather just update it in a consistent manner
+
+    private void retrieveMetadata(){
+        try (MongoClient mongoClient = MongoClients.create(System.getProperty("mongodb.uri"))) {
+            MongoDatabase db = mongoClient.getDatabase("FileSystem");
+            MongoCollection<Document> collection = db.getCollection("Metadata");
+            FindIterable<Document> docs = collection.find();
+
+            for (Document doc : docs) {
+                fileLocator.put((String) doc.get("filename"), (Integer) doc.get("chunk"));
+                loadTracker.put((Integer) doc.get("chunk"), (Integer) doc.get("chunk size"));
+            }
+        }
+    }
+
+    /**
+     * Backs up both the data of the fileLocator and loadTracker maps in the database
+     */
+    private void backupMetadata(){
+        try (MongoClient mongoClient = MongoClients.create(System.getProperty("mongodb.uri"))) {
+            MongoDatabase db = mongoClient.getDatabase("FileSystem");
+            db.getCollection("Metadata").drop();
+            db.createCollection("Metadata");
+            MongoCollection<Document> collection = db.getCollection("Metadata");
+
+            for (Map.Entry entry : fileLocator.entrySet()) {
+                Document data = new Document("_id", new ObjectId());
+                data.append("filename", entry.getKey())
+                    .append("chunk", entry.getValue())
+                    .append("chunk size", loadTracker.get(entry.getValue())) ;
+                collection.insertOne(data);
+            }
+        }
     }
 
     @Override
     public ByteBuffer getFile(String name) throws TException {
-        try (TTransport transport = new TSocket("localhost", 8000)) {
+        if (!fileLocator.containsKey(name)){
+            throw new TApplicationException("File not found");
+        }
+        int chunkPort = fileLocator.get(name);
+
+        try (TTransport transport = new TSocket("localhost", chunkPort)) {
             transport.open();
 
             TProtocol protocol = new TBinaryProtocol(transport);
             chunkClient = new Chunk.Client(protocol);
 
             ByteBuffer buffer = chunkClient.getFile(name);
+            logger.logInfo("REQUEST - GET; CHUNKSERV => " + chunkPort + " - FILE => " + name + " - SIZE => " + buffer.array().length * 1000 + " KB");
 
             Runtime.getRuntime().addShutdownHook(new Thread(() -> {
                 transport.close();
@@ -89,15 +125,22 @@ public class FileSystemImpl implements FileSystem.Iface {
 
     @Override
     public void uploadFile(String name, ByteBuffer file) throws TException {
-        this.logger.logInfo("Name received: " + name);
+        Map.Entry<Integer, Integer> chunkServer = Collections.min(loadTracker.entrySet(), new Comparator<Map.Entry<Integer, Integer>>() {
+            public int compare(Map.Entry<Integer, Integer> entry1, Map.Entry<Integer, Integer> entry2) {
+                return entry1.getValue().compareTo(entry2.getValue());
+            }
+        });
 
-        try (TTransport transport = new TSocket("localhost", 8000)) {
+        try (TTransport transport = new TSocket("localhost", chunkServer.getKey())) {
             transport.open();
-
             TProtocol protocol = new TBinaryProtocol(transport);
             chunkClient = new Chunk.Client(protocol);
-
             chunkClient.uploadFile(name, file);
+
+            logger.logInfo("REQUEST - UPLOAD; CHUNKSERV => " + chunkServer.getKey() + " - FILE => " + name + " - SIZE => " + file.array().length * 1000 + " KB");
+            loadTracker.replace(chunkServer.getKey(), chunkServer.getValue() + 1);
+            fileLocator.put(name, chunkServer.getKey());
+            backupMetadata();
 
             Runtime.getRuntime().addShutdownHook(new Thread(() -> {
                 transport.close();
@@ -109,15 +152,21 @@ public class FileSystemImpl implements FileSystem.Iface {
 
     @Override
     public void updateFile(String name, ByteBuffer file) throws TException {
-        this.logger.logInfo("Name received in update operation: " + name);
+        if (!fileLocator.containsKey(name)){
+            throw new TApplicationException("File not found");
+        }
+        int chunkPort = fileLocator.get(name);
 
-        try (TTransport transport = new TSocket("localhost", 8000)) {
+        try (TTransport transport = new TSocket("localhost", chunkPort)) {
             transport.open();
 
             TProtocol protocol = new TBinaryProtocol(transport);
             chunkClient = new Chunk.Client(protocol);
 
             chunkClient.updateFile(name, file);
+
+            logger.logInfo("REQUEST - UPDATE; CHUNKSERV => " + chunkPort + " - FILE => " + name + " - SIZE => " + file.array().length * 1000 + " KB");
+
 
             Runtime.getRuntime().addShutdownHook(new Thread(() -> {
                 transport.close();
@@ -129,15 +178,24 @@ public class FileSystemImpl implements FileSystem.Iface {
 
     @Override
     public void deleteFile(String name) throws TException {
-        this.logger.logInfo("Name received in delete operation: " + name);
+        if (!fileLocator.containsKey(name)){
+            throw new TApplicationException("File not found");
+        }
+        int chunkPort = fileLocator.get(name);
 
-        try (TTransport transport = new TSocket("localhost", 8000)) {
+        try (TTransport transport = new TSocket("localhost", chunkPort)) {
             transport.open();
 
             TProtocol protocol = new TBinaryProtocol(transport);
             chunkClient = new Chunk.Client(protocol);
 
             chunkClient.deleteFile(name);
+            logger.logInfo("REQUEST - DELETE; CHUNKSERV => " + chunkPort + " - FILE => " + name);
+
+            loadTracker.replace(chunkPort, loadTracker.get(chunkPort) - 1);
+            fileLocator.remove(name);
+            backupMetadata();
+
 
             Runtime.getRuntime().addShutdownHook(new Thread(() -> {
                 transport.close();
